dev-rust/ctrlc: import ctrlc-3.2.1 for crosvm
Add the ctrlc crate so that it can be used in chromeos-base/crosvm.

This is intentionally one minor version behind the latest, as the newest
version bumped the required nix version to 0.24, which we don't have in
the tree yet; picking this version will let us keep a single version of
dev-rust/nix in the tree. (This was the only change in ctrlc-3.2.2, so
it is not a problem to stick with the older version for now).

BUG=b:213146388
TEST=emerge-hatch crosvm

Change-Id: Iba0f02530a605008f7c6bcbdd86c42f9945e5dc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3807928
Commit-Queue: Chih-Yang Hsia <paulhsia@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Tested-by: Chih-Yang Hsia <paulhsia@chromium.org>