sys-apps/flashrom: Disable unused programmers
Disable some programmers because they are unused.

With the recent work on meson improvements, it has been discovered
that some programmers which can only run on x86, are attempted to be
deployed on ARM machines. This situation creates a clash, but since
the programmers are not used the easiest solution is to disable them.

Consequently, the events above inspired the review of all the
programmers enabled by default. We discovered many of them are not
really used and should not be enabled by default in our tree. Which
also makes the binary smaller.

BUG=b:241313027
TEST=two scenarious work as expected
1) nm programmer_table.c.o
before and after the patch shows only enabled programmers
2) console output from emerging flashrom, before and after the patch
lists all meson config options with true/false values for enabled/
disabled programmers

Change-Id: Ied413403e6c0f61098992fcd3c59e874aa0268b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3809071
Commit-Queue: Anastasia Klimchuk <aklm@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>
Tested-by: Anastasia Klimchuk <aklm@chromium.org>
Reviewed-by: Peter Marheine <pmarheine@chromium.org>