skyrim: Add winterhold coreboot configuration
Create a new coreboot configuration for the winterhold variant
of the skyrim reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:240970782
TEST=FW_NAME=winterhold emerge-skyrim coreboot chromeos-bootimage
Ensure that image-winterhold.*.bin are created

Cq-Depend: chromium:3834630
Change-Id: Icc9c8561be9142882a21de6027c27f4a9e305db5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3834051
Commit-Queue: Isaac Lee <isaaclee@google.com>
Tested-by: Isaac Lee <isaaclee@google.com>
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>
Reviewed-by: Raul Rangel <rrangel@chromium.org>