nissa: Add yaviks coreboot configuration
Create a new coreboot configuration for the yaviks variant
of the nissa reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

BUG=b:242277219
TEST=FW_NAME=yaviks emerge-nissa coreboot chromeos-bootimage
Ensure that image-yaviks.*.bin are created

Cq-Depend: chrome-internal:4950126
Change-Id: I6fe9402b2d29f63db764b38ec58812fa70f0ca9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3826621
Reviewed-by: Reka Norman <rekanorman@chromium.org>
Tested-by: Wisley Chen <wisley.chen@quanta.corp-partner.google.com>
Commit-Queue: Wisley Chen <wisley.chen@quanta.corp-partner.google.com>
Reviewed-by: Super Ni <super.ni@intel.corp-partner.google.com>