sys-libs/glibc: make crypt.h install depend on crypt use flag
This commit is the chomiumos-overlay counterpart to the
upstream Gentoo commit which landed yesterday:

https://github.com/gentoo/gentoo/commit/77cec48da70c6d6424ed6dba4357dd8eacd262c2

It depends on the glibc 2.35 uprev because it's required
for building glibc with Clang which makes sense only for
2.35 or later and the uprev will land soon anyway.

The original upstream commit msg:

The crypt use flag is supposed to control whether libcrypt
and its associated crypt.h are installed, but it's ignored
in header-only builds and crypt.h is always installed.

This generates a conflict for eg with sys-libs/libcxrypt
installed as a system lib which provides /usr/include/crypt.h
even if glibc is built with -crypt.

The solution is for glibc to properly respect the crypt
use flag when installing the headers.

BUG=b:241257035
TEST=Local builds; CQ.

Cq-Depend: chromium:3770110
Change-Id: I381f32bbd5732a7b3122e8f41b8e92e16f52f6a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3843374
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>