skyrim: Add morthal coreboot configuration
Create a new coreboot configuration for the morthal variant
of the skyrim reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:240970782
TEST=FW_NAME=morthal emerge-skyrim coreboot chromeos-bootimage
Ensure that image-morthal.*.bin are created

Cq-Depend: chromium:3856394
Change-Id: I9f2b045e76c36662a4aebcb7f918f546476f6156
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3857721
Commit-Queue: Moises Garcia <moisesgarcia@google.com>
Reviewed-by: Isaac Lee <isaaclee@google.com>
Reviewed-by: Jonathon Murphy <jpmurphy@google.com>
Tested-by: Moises Garcia <moisesgarcia@google.com>