vim, vim-core: upgrades packages & move to overlay
Upgraded app-editors/vim to version 9.0.0108 on amd64
Upgraded app-editors/vim-core to version 9.0.0108 on amd64
Adds the vim and vim-core packages to chromiumos-overlay because the
latest stable version on Gentoo has a cross-compilation bug and cannot
be utilized as is. This version pulls in a more up to date tagged
version. b/241808816 tracks the issue.
Replaces a couple of banned keywords in the comments inside the
vim-core/files/vimrc-r6 file.

BUG=chromium:1335706, chromium:1336938, chromium:1337542,
chromium:1340924, chromium:1343339, chromium:1344118, chromium:1344821,
chromium:1346256
TEST=Both vim and vim-core compile correctly for SDK and amd64-generic.
TEST=`vim --version` shows the correct version.

Change-Id: I44d6a30a3165b2a81974509f85dfe0b67f38198d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3846420
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>