cros-rust: fix a race condition with registry ownership
In migrating packages to third-party-rust-crates, I'm occasionally
observing the following sequence of events at chroot creation:

```
>>> Emerging dev-rust/byteorder
>>> Installing dev-rust/byteorder
>>> Completed dev-rust/byteorder
[ Many more packages are installed ]
>>> Emerging dev-rust/third-party-crates-src
>>> Installing dev-rust/third-party-crates-src
>>> Completed dev-rust/third-party-crates-src
>>> Uninstalling dev-rust/byteorder
[ More packages are installed ]
>>> Completed dev-rust/byteorder
```

third-party-crates-src has a soft blocker on byteorder, since it
replaces the sources used by byteorder. It seems the uninstallation of
byteorder is unsequenced with the installation of any package which
depends on third-party-crates-src, which presents a window for a race.
In the happy path:
1. Byteorder's prerm removes the symlink that third-party-crates-src
   installed.
2. Byteorder's postrm readds the symlink that third-party-crates-src
   installed.
3. A package which depends on said symlink through depending on
   third-party-crates-src attempts to use byteorder

Unfortunately, this sequence can happen in the order `1, 3, 2`, which
causes Cargo to complain about `byteorder` not existing in the registry.

This CL hopes to make the above not happen by removing our `postrm` hook
wholesale, and only allowing `prerm` to remove a registry link if the
package trying to `prerm` the link is the package that added the link in
the first place.

BUG=b:240953811
TEST=Set up a chroot 5x; no races noted.

Change-Id: I5e77a9dc7f0c7ebd752401aaa7b9bb6681ea60bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3856975
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>