sys-libs/glibc: add llvm use flag + patches
This adds the latest patches from Adhemerval's tree to
build glibc with an LLVM/Clang based toolchain.

libgcc is used as a compiler runtime to avoid a
bootstrapping problem as documented in b:241257035.

The patches and all the Clang related logic logic are
behind an "llvm" use flag to allow for testing, but
there are still some rough edges which should be fixed,
like the removal of the -mthumb flag in favor of -marm.

The llvm use flag is enabled in the default packages.use
to allow for CQ testing but this should definitely NOT
be enabled in production.

BUG=b:237545941
TEST=Local builds & boot on x64, arm64; CQ.

Cq-Depend: chromium:3770110, chromium:3843374, chromium:3892354
Change-Id: Iee631b3a02543962cd7d74e722fc11ae968a3ced
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3858262
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>