shill: Add missing dependencies
Since shill no longer depends on vpn-manager, it should has dependency
on xl2tpd and strongswan directly to install those packages. This patch
adds these two packages in the RDEPEND field in the ebuild file of
shill.

Note that vpn-manager also depends on ppp, which is already in the
dependency list of shill.

BUG=b:245000912
TEST=Unemerged strongswan and xl2tpd package, checked that these two
  packages will not be installed when emerging shill before this patch,
  and will be installed after this patch.

Change-Id: I608d8d9933aa47cb9c8c65064a609f356d4edf6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3874247
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>