dev-embedded/hps-sdk: update Rust to 1.62.1
Patches are taken as is from dev-lang/rust.

The copy of miri bundled with rustc under src/tools/miri does not
compile, so we replace it with a slightly later version of miri from its
upstream git repo, after it had been updated to account for all the
compiler changes in 1.62.

BUG=b:241835221
TEST=sudo emerge hps-sdk && emerge-brya hps-firmware{,-tools}
TEST=scripts/run-tests
TEST=cargo miri setup && scripts/run-tests-miri
TEST=scripts/test-models

Change-Id: I585d56ca44530852057f91169071bc0b74f6e25d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3878402
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>