cros-rust: fix bug with ownership
When attempting to add a chroot update hook, I see:
```
/var/tmp/portage/._unmerge_/dev-rust/strsim-0.8.0-r1/temp/environment:
line 509:
/usr/lib/cros_rust_registry/owners//usr/lib/cros_rust_registry/registry/strsim-0.8.0:
No such file or directory
 * strsim-0.8.0 removal from Cargo registry skipped due to new symlink
   owner
```

This means that old registry symlinks won't be removed properly. Cargo
does not seem to care about dead symlinks in the registry, and we remove
symlinks appropriately in `src_preinst`. Given that we're replacing most
things with dev-rust/third-party-crates-src, this shouldn't be much a
problem since it was caught quickly.

BUG=b:240953811
TEST=sudo emerge anyhow && sudo emerge -C anyhow

Change-Id: Ieb8064c5e06faa32546c03353e188b1153707536
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3893030
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>