syzkaller: patch to introduce /rawcover32
syz-manager provides the /rawcover API that allows for retrieving
covered PCs. This requires that the fuzzing configuration specify
"kernel_obj".

When syzkaller is run via the tast wrapper, it does not have access to
build artifacts such as vmlinux, and /rawcover is not usable.

Introduce the /rawcover32 API that allows for retrieving PCs without the
"kernel_obj".

BUG=b:245574620
TEST=emerge-brya --root=/tmp/m dev-go/syzkaller
TEST=emerge-octopus --root=/tmp/m dev-go/syzkaller
TEST=emerge-trogdor --root=/tmp/m dev-go/syzkaller
TEST=emerge-trogdor64 --root=/tmp/m dev-go/syzkaller

Change-Id: I423e9e57d329e782ac2e7143003cfd85237f4eef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3895414
Commit-Queue: Zubin Mithra <zsm@chromium.org>
Tested-by: Zubin Mithra <zsm@chromium.org>
Reviewed-by: Martin Faltesek <mfaltesek@google.com>