perf: Upgrade to 5.15.68.
5.15.0 is missing commits like 05ef4f56173e ("perf evlist: Fix failed to
use cpu list for uncore events"), necessary to use commands like
`perf stat -C0 -e msr/tsc/`, which are used for several performance
tests. Update to 5.15.68 to pick these up.

Drop "perf session: Remap buf if there is no space for event", since it
is included in 5.15.68.

Add -lopencsd to LDFLAGS when use opencsd to work around 115b85125624
("perf tools: Drop requirement for libstdc++.so for libopencsd check"),
which expects -static to be included in LDFLAGS if opencsd_c_api is a
static library; this is not the case for opencsd_c_api, so the -lopencsd
is needed.

BUG=b:226008810
TEST=`perf stat -C0 -e msr/tsc/ sleep 1` on nirwen

Change-Id: I2fa93db6450d53f31bcc32bf5d5a888064c22156
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3901805
Reviewed-by: Victor Ding <victording@chromium.org>
Tested-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>