sys-libs/scudo: Add initial scudo ebuild
Emerging this package enables scudo by default on the
target board. This is based on the compiler-rt ebuild.

This commit does not introduce any of the necessary scudo
patches for ChromeOS usage.

This package should not be built into the image at this
time, as it requires resource limits to be modified/removed
and seccomp filter changes. The device will not be able
to get a network connection until those changes are in
effect.

BUG=b:225008839
TEST=emerge-volteer sys-libs/scudo
TEST=emerge-jacuzzi sys-libs/scudo
TEST=emerge-trogdor64 sys-libs/scudo

Change-Id: I60f29cc1a067e211f6909995f3e2f3156a3fdb34
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3908097
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>