Sabrina: Remove embargoed name
'Mendocino' was an embargoed name and could previously not be used
in references to Skyrim.  Coreboot has references to sabrina both
in directory structure and in files. This will make life difficult
for people looking for Mendocino support in the long term. The code
name should be replaced with "mendocino".

BUG=b:239072117
TEST=Builds

Cq-Depend: chrome-internal:4878294
Cq-Depend: chromium:3764023
Cq-Depend: chromium:3914916
Change-Id: I7dad16761d15c1cbc8a197101bd8e09aac994be9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3763392
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Tested-by: Jonathon Murphy <jpmurphy@google.com>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>