scudo: Add ChromeOS config to llvm scudo
This adds a local out-of-tree LLVM patch to add
ChromeOS platform config to Scudo. Notably,
the important change here is the reduction of the
PrimaryRegionSizeLog from 32 to 24, which ensures
that we don't mmap any larger than (in this case)
44*2^24 bytes, or 704MiB for the primary allocator.

This allows us to be under some important address
space limits for certain daemons which are managed
by upstart. This however is still insufficient to
run some network-related daemons which are needed
for SSH.

BUG=b:225008839
TEST=emerge-volteer sys-libs/scudo (check defines are set)

Change-Id: I3d119b26194eeb144884cf7eecf0833887d89622
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3916671
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>