avahi-daemon: Disable rlimits
Remove conf lines that explicitly set the rlimits for the data segment
and process stack, which has the effect of leaving the system defaults.
This workaround has been reported as a potential fix
for avahi OOMs, though this is dated info. Frankly, no idea if this will work but it seems like it's worth trying out.

BUG=b:244521026
TEST=build & flash brya

Change-Id: I8a03be41f34f29b224a7654a168ebd43a4c5cc44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3924594
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>