perfetto: generate the SDK source by default
This change enables crrev/c/3908039 in the stable ebuild for generating
the SDK source instead of relying on the pre-generated SDK source. This
change also bumps the Perfetto commit to include aosp/2223046 for the
SDK generation script to use the system buildtools.

BUG=b:247942837
TEST=build_packages --board $BOARD
TEST=CQ dry run passes

Cq-Depend: chromium:3924884
Change-Id: Iac069a9d91daf2e86d4a8b07954aa65352ddacca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3929722
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Chinglin Yu <chinglinyu@chromium.org>
Commit-Queue: Chinglin Yu <chinglinyu@chromium.org>