avahi-daemon: Ignore non-physical interfaces.
This patch updates the deny-interfaces stanza by replacing br0 (the name
of the very old arc_networkd bridge) with a list of interfaces that can
be created for connected namespaces and guest VMs.

This doesn't appear to have any impact to avahi's cpu usage in the
crosfleet lab but it does seem to eliminate/reduce extra work for
multicast forwarder.

It may have a positive impact of memory usage though this is yet to be
confirmed - but the crosfleet redrix, while returning to 20+% cpu has
yet to exceed 10MB vmpeak.

BUG=b:244105139
BUG=b:244521026
TEST=deployed change to 2 bryas on test vlan, ran mdns discover app to
verify can still see devices; ran spotify app on both devices to verify
discover and broadcasting works.
TEST=deployed to crosfleet brya, observed patchpanel's mcast forwarder
drops off the list of top cpu consumers.

Change-Id: Ib36f90db25aa9c4ef9c1b9b6e2ed044ed05c155f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3932394
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>