coreboot-utils: Install ifdtool on the device as well as the host
Nissa is using a new factory flow which involves using ifdtool to lock
the Intel descriptor in factory. So install ifdtool on the device so it
can be used in factory. It's required in the base image as well as the
factory image in order to support shimless RMA.

BUG=b:255462682
TEST=`emerge-nissa coreboot-utils && cros deploy <dut> coreboot-utils`,
check ifdtool is present on the device and works.

Cq-Depend: chromium:4096701
Change-Id: I99f99447855e559e2a4b47eca9dc4c1d5334b6b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4018976
Reviewed-by: Sam McNally <sammc@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>
Commit-Queue: Reka Norman <rekanorman@chromium.org>