perfetto: force disable PERFETTO_DCHECK() in the SDK
This change explicitly disable PERFETTO_DCHECK() in the SDK header by
adding perfetto_force_dcheck="off" to the SDK GN args. This avoids the
inconsistency of PERFETTO_DCHECK_IS_ON() between the SDK binary and the
embedder, which causes a runtime crash to the embedder.

The SDK should be generated so that the header behaves the same
regardless of -DNDEBUG, but some embedder can enable DCHECK_ALWAYS_ON
to force enable PERFETTO_DCHECK() in the header. Setting
perfetto_force_dcheck to "off" rules out this possibility.

BUG=b:245197717
TEST=emerge-$BOARD perfetto and check PERFETTO_FORCE_DCHECK_OFF() in the
     SDK header.

Change-Id: Icd75f837513d27dcd6cf3d5c90077140a7628602
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3936105
Tested-by: Chinglin Yu <chinglinyu@chromium.org>
Reviewed-by: Primiano Tucci <primiano@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Sami Kyöstilä <skyostil@chromium.org>