net-dns/avahi-daemon: Reduce cache size
This patch changes the cache entry limit to 100 records from the default
4096 which may be suitable for machines on large LANs actively wanting
to participate in local discovery but is too high for the average user.
The impact of this change is that it reduces the amount of time avahi
spends walking the cache whenever it handles a query packet due to the
expensive key comparator.

Testing on crosfleet redrix machines showed an obvious and repeatable
reduction in CPU consumption by half or more, with control machines in
the 15-25% range and treatment machines in the 5-10% range.

Eliminating caching altogether does not appear to introduce further
benfit and may have an adverse impact by causing more traffic by
reiussing queries.

BUG=b:244105139
TEST=deployed to crossfleet redrix, ran for 6 hours (see above)

Change-Id: I554ae6cc10983c50decefc574360c29fcaa94f44
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3937062
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Tested-by: Garrick Evans <garrick@chromium.org>