migration_utils: allow soft blockers in RDEPEND
A handful of crates are considered non-leaves because they have
soft-blockers in their RDEPEND. Ignore those.

This may seem unsafe for folks with old chroots. It shouldn't be a
problem in practice, since we recently `emerge -C dev-rust/*`'ed as part
of chroot upgrades.

BUG=b:240953811
TEST=CQ; local builds of atlas, trogdor; chroot creation

Change-Id: I72d1ce13751e1940170c14aa900657f6f2ad15b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3938434
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>