skyrim: Add frostflow coreboot configuration
Create a new coreboot configuration for the frostflow variant
of the skyrim reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:240970782
TEST=FW_NAME=frostflow emerge-skyrim coreboot chromeos-bootimage
Ensure that image-frostflow.*.bin are created

Cq-Depend: chromium:3941093
Change-Id: I21d59cd17cb46269d5a0f3cdb67bc19e6cbd5c4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3939114
Tested-by: Chao Gui <chaogui@google.com>
Commit-Queue: Chao Gui <chaogui@google.com>
Reviewed-by: Bob Moragues <moragues@chromium.org>
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@google.com>