secagentd: Enable by default
Enable the building and installation of secagentd for platforms that
have a 5.10+ kernel by default. The upstart config for secagentd is
purposely not installed to prevent it from being started on platforms.

The purpose of this change is to ensure that CQ catches build related
errors before the feature is fully launched.

BUG=b:244752495
TEST=emerge-amd64-generic target-chromium-os after changing kernel to
4.4, 4.14, 5.4. Verified secagentd was only ever installed on 5.10
  kernels and later.

Cq-Depend: chromium:3964867
Change-Id: Icdef0672296a757430187384bdffa9c1960dc8c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3946374
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Auto-Submit: Jason Ling <jasonling@chromium.org>
Tested-by: Jason Ling <jasonling@chromium.org>