herobrine: Add zombie coreboot configuration
Create a new coreboot configuration for the zombie variant
of the herobrine reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:249180463
TEST=FW_NAME=zombie emerge-herobrine coreboot chromeos-bootimage
Ensure that image-zombie.*.bin are created

Change-Id: I85b2f8ffce6687d42aa2596b8173cfe1fab4a94f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3946378
Reviewed-by: Bob Moragues <moragues@chromium.org>
Tested-by: Bob Moragues <moragues@chromium.org>
Reviewed-by: Shelley Chen <shchen@chromium.org>
Commit-Queue: Bob Moragues <moragues@chromium.org>