wpa_supplicant: fix lint errors in ebuild
It seems that it was a tradition to upload changes to supplicant ebuild
with "--no-verify".  Let's be rebellious and break that tradition and
fix lint errors.

Some of the changes were not really necessary - extending double quote
to whole file name: "${FILESDIR}"/... -> "${FILESDIR}/..."
but were performed for consistency.

BUG=None
TEST=emerge-$BOARD wpa_supplicant-cros
TEST=Run "FEATURES=noclean emerge-$BOARD wpa_supplicant-cros" before and
  after this change and check that the generated .config is the same.
TEST=repo upload --verify --dry-run

Change-Id: I631f071adf67a1fe92edbb1d1111f99f0b87a6ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3952171
Tested-by: Andrzej Ostruszka <andrzejo@google.com>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>
Reviewed-by: Sergey Frolov <sfrolov@google.com>