Prepare shared crosvm crates for uprev.
Use the same versioning scheme as for dev-rust/sync to avoid cargo build
errors when both the crosvm and ChromeOS versions of these crates are
included in the same build at the same time.

Rename base to crosvm-base when it is included in these packages to use
the ChromeOS version.

BUG=b:250883877
TEST=~/chromiumos/chromite/scripts/cros_uprev --force --overlay-type \
  public --packages \
  dev-rust/sync:dev-rust/io_uring:dev-rust/cros_async:chromeos-base/crosvm-base:dev-rust/data_model:dev-rust/serde_keyvalue:dev-rust/serde_keyvalue_derive

Change-Id: I9e637a7ee53fb8371b86f10022653ccf9e4e29b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3954344
Tested-by: Allen Webb <allenwebb@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>
Commit-Queue: Allen Webb <allenwebb@google.com>