profiles: base: move code coverage reports generated by tast-cmd
When 'cros_run_unit_tests' is run with the 'USE=coverage', it will now
generate code coverage reports for 'tast-cmd' in a Portage temporary directory. In order to access the code coverage, the files have to be moved to a different directory based on this documentation: https://chromium.googlesource.com/chromiumos/docs/+/HEAD/portage/ebuild_faq.md#how-do-i-store-transient-artifacts-in-an-ebuild. Corresponds with change to chromeos-base/tast-cmd/tast-cmd-9999.ebuild.

BUG=b:247635178
TEST=Ran 'cros_run_unit_tests' with and without the coverage flag

Change-Id: Idcb66a7d280ce8c67c8dd03cb14d0e28ca67dad5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3936766
Tested-by: Charles Liu <liucharles@chromium.org>
Commit-Queue: Charles Liu <liucharles@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Srinivas Hegde <srinivashegde@google.com>