brya: Add marasov coreboot configuration
Create a new coreboot configuration for the marasov variant
of the brya0 reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:254365935
TEST=FW_NAME=marasov emerge-brya coreboot chromeos-bootimage
Ensure that image-marasov.*.bin are created

Change-Id: I1b1cf0f2d5e56e2d59cef3ed3d7be2ee13df8dc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3977748
Tested-by: Frank Chu <frank_chu@pegatron.corp-partner.google.com>
Commit-Queue: Kyle Lin <kylelinck@google.com>
Reviewed-by: Kyle Lin <kylelinck@google.com>