floss: Add new instance param in btadapterd: INDEX
btadapterd will now have two instance params: index and hci. The index
is the virtual index that will be used in apis and the hci is the
physical hci index. In btmanagerd, the two will be kept in sync based on
the devpath for the emitted physical hci index.

BUG=b:244457509
TEST=Suspend/resume Realtek and confirm INDEX=0

Cq-Depend: chromium:3976884
Change-Id: Icb386ff604d6cfe57199ae251bed3177372c56ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3975854
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>