media-libs/tiff: Move utilities to /usr/local
In addition to the main shared library, tiff ships with some
command-line utility programs.  These programs are not used by anything
in a regular ChromeOS image, but they have been the source of some
security issues in the past.  Move the utility programs to /usr/local to
keep them available for debugging while reducing the possibility of
hitting those code paths on a non-test image.  This also saves ~350KB on
the rootfs.

BUG=chromium:1377610, chromium:1359678
TEST=tast run '("group:paper-io")'

Change-Id: I745734f6ddf5ac08ad87516715e33513ab256f93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3987226
Reviewed-by: Piotr Pawliczek <pawliczek@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>