prjoxide: Patch LRAM timing to be for registered outputs
The timing for non-registered outputs is around 12ns which is way too
high to meet timing at a satisfactory clock frequency. Ideally prjoxide
and nextpnr would be changed to have correct output timing for both
registered and non-registered outputs and select based on the actual
setting used. Doing this is left as an exercise for the future.

BUG=b:243330750
TEST=sudo emerge prjoxide && sudo emerge nextpnr
TEST=cd hps-firmware2; ./scripts/build-fpga-bitstream &&
./scripts/test-models

Cq-Depend: chromium:3960659, chromium:3960658, chromium:3976828
Change-Id: I830d1edb5b2067f1351ed069918ad86578eb54b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3989524
Reviewed-by: Dan Callaghan <dcallagh@chromium.org>
Commit-Queue: David Lattimore <dml@chromium.org>
Tested-by: David Lattimore <dml@chromium.org>