featured: Update platform-features.json permission
Currently platform-features does not have write permissions, which is
causing several build failures in Lacros.

Update permissions to include write. This is a speculative fix; we believe including write permissions will resolve the permission denied errors but testing this is time consuming. We are working to test this but are reverting to unblock existing builds. We will use pre-submits to ensure this fix does not break any existing code.

BUG=b:256012263
TEST=None

Change-Id: Ib76bf36161ec489487911095e86da1688c67e337
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/3994432
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Kendrake Tsui <kendraketsui@google.com>
Tested-by: Kendrake Tsui <kendraketsui@google.com>