eclass/cros-workon: Only set vcsid if get_rev is successful
Right now we are setting an invalid VCSID if the path doesn't contain
a git repo. This will cause anything that uses librillo to fail to
compile because it validates the VCSID format.

When we build with bazel, we don't inject the .git directory, so we
can't extract the VCSID.

Original-Change-Id: Ia761aee2a2e18d171899663cb8aa0454d24bedd7
Original-Reviewed-on: https://team-review.git.corp.google.com/c/cros-build-tiger/third_party/chromiumos-overlay/+/1626050
Original-Reviewed-by: Shuhei Takahashi <nya@google.com>

BUG=b:262460834
TEST=BOARD=arm64-generic bazel build --//bazel/sdk:new-sdk @portage//virtual/target-os:all

Change-Id: I4cf7d7184c7c99c66f33e87dcb0cdf93cf197fa6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4371244
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>