Rust upgrade to 1.65.0.
llvm-project is now symlinked within the Rust source directory, which is
faster than copying.

New cherry picked patches for rust-llvm-sources that we will be able to
remove once we are on a newer LLVM:

ARM-Add-target-feature-to-force-32-bit-atomics.patch
DwarfEhPrepare-Assign-dummy-debug-location-for-inser.patch
InlineAsm-Improve-error-messages-for-invalid-constra.patch

I made the new rust-llvm-sources patch
Remove-include-in-int_util.c.patch
because the `sed` command we previously used in `cros-rustc.eclass` is
apparently no longer sufficient now that llvm-project is symlinked.

I moved the patch
Revert-CMake-Unconditionally-add-.h-and.td-files-to.patch
from a Rust patch (patching the internal llvm-project) to a
rust-llvm-sources patch, where it makes more sense.

Removed these patches which are now incorporated into upstream:
rust-Bump-cc-version-in-bootstrap-to-fix-build-of-uefi-ta.patch
rust-Don-t-build-std-for-uefi-targets.patch
rust-passes-only-in-pre-link.patch

Removed this patch which is now obsolete:
rust-Remove-DXILPointerTyID-which-our-LLVM-doesn-t-have.patch

Made this new Rust patch which we'll need temporarily until we move to
a newer LLVM:
rust-Remove-LLVMGetAggregateElement.patch

Removed the patch
rust-no-test-on-build.patch
This patch was previously used because our build was configured to build
`library/std`, which was impossible for some targets, so we needed to
patch the build not to actually build std for those targets. It turns
out the right thing to do is to build the target `library`, which will
build `std` for targets where that is appropriate and only `core` for
targets where that is appropriate. I don't think this behavior is
documented.

BUG=b:257261436
TEST=CQ

Cq-Depend: chromium:4051348
Change-Id: I5122e5d278dd059f942131acfee15c63cbe03cde
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4037547
Tested-by: Michael Benfield <mbenfield@google.com>
Reviewed-by: George Burgess <gbiv@chromium.org>