chromeos-test-root: Add fake camera HAL to test images
Add fake camera HAL to test image, which reads config and emulates
cameras for testing purpose.

This is only needed for test image, but needs to live in the rootfs
since the camera service only searches and loads HAL .so files in
/usr/{lib,lib64}/camera_hal.

The fake.so file currently is ~175KB on brya (x86_64) device.

BUG=b:236782829
TEST=build_image --board brya test, flash to DUT and check
     /usr/lib64/camera_hal/fake.so exists.

Change-Id: I7f05f4067be4128b561750c4733b7a5cd4717be8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4021874
Commit-Queue: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>