assistant-dlc: Add USE flag constraint on chromeos-chrome dependency.
assistant-dlc will fail if libassistant.so does not exist. It comes
from a copy of chromeos-chrome built with chrome_internal, so enforce
this requirement. A consequence of this change is Chrome may be rebuilt
or re-installed if a non-internal build was previously present.

BUG=b:257155890
TEST=./build_packages

Change-Id: I1e1154fea9b5697955eed475c9b8e48b372f8e1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4032025
Auto-Submit: Allen Webb <allenwebb@google.com>
Tested-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Allen Webb <allenwebb@google.com>
Reviewed-by: Tao Wu <wutao@chromium.org>