optee_os: Extract ta_dev_kit creation from the optee_os ebuild
Longer term, the plan is to include all TAs with the OpTee image as
"early TAs". That means we need to build OpTee in 3 stages:

1. Build TA dev kit to allow building of TAs
2. Build TAs
3. Build OpTee and include TAs in resulting image

This breaks that first step out to be separate to allow the 3 steps to
occur (building TAs was already separate).

BUG=b:248609772
TEST=emerge-cherry optee-oemcrypto-ta (also builds optee_os_tadevkit)

Cq-Depend: chrome-internal:5106514
Change-Id: I0b3466adb2284a9543099fd014d93120ee06a070
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4035357
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>