cros-ish: Disable Kconfig checker from ebuild
The Kconfig checker is being executed for cros-ish builds when it
shouldn't.

From https://crrev.com/c/3450757:
"...it has also recently been responsible for multiple blow-ups,
and as such, we should limit the scope to where it runs to the CLs
that might actually affect it.  The CLs that land in platform/ec do
not use this ebuild, and therefore this is a spot we should disable
it."

BUG=b:181323955
TEST=emerge-drallion chromeos-ish
TEST=emerge-sarien chromeos-ish

Cq-Depend: chromium:4067442
Change-Id: If0c55aa900a2549d1be096e81211b5d3f6eaa7cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4067356
Tested-by: Mark Hasemeyer <markhas@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Mark Hasemeyer <markhas@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>