accessibility: Update pumpkin DLC to version 3.0
We recently added more semantic tags to the Pumpkin
DLC [1], which allows it to recognize more commands.
We uploaded a new version of the DLC to Google
Cloud Storage; this change updates the Pumpkin
ebuild so that it pulls down the latest version.

[1] https://critique.corp.google.com/cl/492278572

BUG=b:1258190
TEST=Manually on device

Test steps:
1. (chroot) emerge-eve app-accessibility/pumpkin
2. (chroot) du -sh /build/eve/build/rootfs/dlc/pumpkin/package/root/
3. Verify the new size is 5.9MB (the old DLC was 5.1MB).

Change-Id: If47b0db0847424d7435dff3b51c63f50a7a9c5b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4071910
Reviewed-by: Katie Dektar <katie@chromium.org>
Tested-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>