chromeos-base: move udev keyboard rules here
chromeos-base package contains all non-board-specific udev rules that
are needed on ChromeOS. Adding keyboard rules to the udev package itself
was a mistake. Move it here.

Update the license from GPLv2 to our standard one now that the file is
not attached to the udev package.

BUG=None
TEST=Check brya build, CQ

Change-Id: I85fa59c3f8318b43218a744387929d61f2e647a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4087606
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>