chromeos-base: sysctl: Specify "sched_rt_runtime_us" here again
In <https://crrev.com/c/2365278> we commented out the setting of
"sched_rt_runtime_us" in 00-sysctl.conf because we had an init script
(rt-limits.conf) that set this up along with a whole pile of related
settings. All this pile of related settings was needed because we were
using RT_GROUP_SCHED.

It turns out that using RT_GROUP_SCHED wasn't such a great idea. We're
finally turning it off. As such, we don't need a whole init script to
configure these two values. We'll delete the init script in
<https://crrev.com/c/4097630> (which we Cq-Depend on) and then this CL
moves the setting back to 00-sysctl.conf.

Though the limit was (800 ms / 1000 ms) before we started using
RT_GROUP_SCHED in the first place, we'll pick (850 ms / 1000 ms)
here. Though there's no data showing that we are anywhere near RT
throttling with 800 ms, we've seen how badly the system behaves if we
ever have RT throttling and so we'll err slightly more on the side of
giving more RT. The leftover (150 ms / 1000 ms) should be enough to at
least maintain control of a system in the case of a runaway RT
process.

Although it's logically distinct, we'll also use this CL to delete the
snippet of code in the "chromeos-init" ebuild that messed with the
rt-limits.conf, essentially undoing <https://crrev.com/c/2181859>. We
squash this into the same CL just to avoid some extra dependency
chains. In order to keep tests passing we pretty much have to land all
these changes atomically so having one fewer CL in the mix seems like
a good thing.

BUG=b:259314614
TEST=limits.go

Cq-Depend: chromium:4097630
Cq-Depend: chromium:4098164
Change-Id: I416882e36e8367912dccadc5ea16deb6d51f9959
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/4100123
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Joel Fernandes <joelaf@google.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Joel Fernandes <joelaf@google.com>
Reviewed-by: Joel Fernandes <joelaf@google.com>